requirements.txt
django==6.0.3
mysqlclient==2.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
Pillow==12.2.0
python-dotenv==1.2.2
